125.2 Conditional Rendering(Section 11: Let's Test Your React Mastery! ) So in order to fix this we can add in a very simple check inside the video detail inside of your all
say if there is no video if video does not exist then I want to return a div that has the text loading

like so.

So if we don't pay if we pass in a value of NULL for the video Propp we're going to instead show the

divx loading and we're not going to even get down here to this line of code where we attempt to reference

the snippet property on it.

All right.

So let's now test this out.

And when I flip back over I see the text that says loading right here.

Now obviously showing the text loading on the screen without ever having it resolve into something without

user interaction is not the best experience.

But we're going to fix this up and improve this experience in just a little bit.

For right now let's just try searching for something.

It's all say buildings now I should be able to click something and see the title appear here instead

of the text loading.

So I click on the video and I see the title appear perfect and I can click on any of these and all the

appropriate titles show up if I search for some other term like let's say I don't know flowers not what

I expected but ok I can now click on these and I see the appropriate titles appear again.

All right so it looks pretty good to take a quick pause right here.

When we come back the next section we're going to start improving the styling of our video detail just

a little bit.