133.2. Building and Styling the Accordion(Section 12: Understanding Hooks in React)
To get that to go away, we essentially need to not return in each Timal element right here.

You might recall that to do that, we can return a react fragment.

We still have to re return everything, because once we eventually have to dids, we just don't want

to have any extra element right there to supply a react fragment.

Remember, we can delete the element name altogether.

But in this case, we also want to supply the key property.

In this case, we have to write out the full name of react pregnant.

So put in, react thought, fragment like.

So I put in the appropriate closing element at the bottom.

Now, on this wrist, we're seeing this list is rendered onto the screen.

There will be no extra element right here.

This is just a containing j a sex element that react understands.

We've got to dibs inside of it.

So if we save this now and go back over, you'll now see that we do not have that extra border at the

very top.

I can go back over to my app and remove that B.R..

All right.

So looks good.

Well, I'd say that stylistically, everything is coming together pretty well.

So now we just really need to make sure that we've got some functionality where we can click on these

things and collapse them or expand them.

That's where the state system and specifically the hook system is going to come into play.

Let's start to understand the whole system in function components in the next video.